requirements.txt
pandas
streamlit
matplotlib
seaborn
langchain-google-genai
langchain-experimental
pygwalker
python-dotenv
tabulate

-e .